Allow users to set profile pictures
