+++ src/Components/Users/UserProfile.tsx
@@ -1,5 +1,5 @@
 import { useState, useReducer, FormEvent } from "react";
-import { GENDER_TYPES } from "../../Common/constants";
+import { GENDER_TYPES, LocalStorageKeys } from "../../Common/constants";
 import { validateEmailAddress } from "../../Common/validation";
 import * as Notification from "../../Utils/Notifications.js";
 import LanguageSelector from "../../Components/Common/LanguageSelector";
@@ -9,9 +9,11 @@ import {
   classNames,
   dateQueryString,
   formatDate,
+  formatDisplayName,
   isValidUrl,
   parsePhoneNumber,
-} from "../../Utils/utils";
+  sleep,
+} from "@/Utils/utils";
 import CareIcon from "../../CAREUI/icons/CareIcon";
 import PhoneNumberFormField from "../Form/FormFields/PhoneNumberFormField";
 import { FieldChangeEvent } from "../Form/FormFields/Utils";
@@ -27,7 +29,13 @@ import request from "../../Utils/request/request";
 import DateFormField from "../Form/FormFields/DateFormField";
 import { validateRule } from "./UserAdd";
 import { useTranslation } from "react-i18next";
+import { EditableAvatar } from "@/Components/Common/Avatar";
+import Page from "@/Components/Common/components/Page";
 import Loading from "@/Components/Common/Loading";
+import AvatarEditModal from "@/Components/Common/AvatarEditModal";
+import uploadFile from "@/Utils/request/uploadFile";
+import careConfig from "@careConfig";
+
 type EditForm = {
   firstName: string;
   lastName: string;
@@ -111,8 +119,9 @@ const editFormReducer = (state: State, action: Action) => {
 
 export default function UserProfile() {
   const { t } = useTranslation();
-  const { signOut } = useAuthContext();
+  const { signOut, refetchUser } = useAuthContext();
   const [states, dispatch] = useReducer(editFormReducer, initialState);
+  const [editAvatar, setEditAvatar] = useState(false);
   const [updateStatus, setUpdateStatus] = useState({
     isChecking: false,
     isUpdateAvailable: false,
@@ -462,490 +471,546 @@ export default function UserProfile() {
       });
     }
   };
+
+  const handleAvatarUpload = async (file: File, onError: () => void) => {
+    const formData = new FormData();
+    formData.append("profile_picture", file);
+    const url = `${careConfig.apiUrl}/api/v1/users/${authUser.username}/profile_picture/`;
+
+    uploadFile(
+      url,
+      formData,
+      "POST",
+      {
+        Authorization:
+          "Bearer " + localStorage.getItem(LocalStorageKeys.accessToken),
+      },
+      async (xhr: XMLHttpRequest) => {
+        if (xhr.status === 200) {
+          await sleep(1000);
+          refetchUser();
+          Notification.Success({ msg: "Profile picture updated." });
+          setEditAvatar(false);
+        }
+      },
+      null,
+      () => {
+        onError();
+      },
+    );
+  };
+
+  const handleAvatarDelete = async (onError: () => void) => {
+    const { res } = await request(routes.deleteProfilePicture, {
+      pathParams: { username: authUser.username },
+    });
+    if (res?.ok) {
+      Notification.Success({ msg: "Profile picture deleted" });
+      await refetchUser();
+      setEditAvatar(false);
+    } else {
+      onError();
+    }
+  };
+
   return (
-    <div>
-      <div className="p-10 lg:p-16">
-        <div className="lg:grid lg:grid-cols-3 lg:gap-6">
-          <div className="lg:col-span-1">
-            <div className="px-4 sm:px-0">
-              <h3 className="text-lg font-medium leading-6 text-secondary-900">
-                {t("personal_information")}
+    <Page
+      title={t("personal_information")}
+      focusOnLoad={true}
+      backUrl="/facility"
+    >
+      <AvatarEditModal
+        title={t("edit_avatar")}
+        open={editAvatar}
+        imageUrl={authUser?.read_profile_picture_url}
+        handleUpload={handleAvatarUpload}
+        handleDelete={handleAvatarDelete}
+        onClose={() => setEditAvatar(false)}
+      />
+      <div className="lg:grid lg:grid-cols-3 lg:gap-6">
+        <div className="lg:col-span-1">
+          <p className="my-1 text-sm leading-5 text-secondary-600">
+            {t("local_body")}, {t("district")}, {t("state")}{" "}
+            {t("are_non_editable_fields")}.
+          </p>
+          <div className="my-4 flex items-center">
+            <EditableAvatar
+              id="user-profile-picture"
+              imageUrl={authUser?.read_profile_picture_url}
+              name={formatDisplayName(authUser)}
+              onClick={() => setEditAvatar(!editAvatar)}
+              className="h-20 w-20"
+            />
+            <div className="my-4 ml-4">
+              <h3 className="text-lg font-medium leading-6 text-gray-900">
+                {authUser?.first_name} {authUser?.last_name}
               </h3>
-              <p className="my-1 text-sm leading-5 text-secondary-600">
-                {t("local_body")}, {t("district")}, {t("state")}{" "}
-                {t("are_non_editable_fields")}.
+              <p className="text-sm leading-5 text-gray-500">
+                @{authUser?.username}
               </p>
-              <div className="flex flex-col gap-2">
-                <ButtonV2
-                  onClick={(_) => setShowEdit(!showEdit)}
-                  type="button"
-                  id="edit-cancel-profile-button"
-                >
-                  {showEdit ? t("cancel") : t("edit_user_profile")}
-                </ButtonV2>
-                <ButtonV2 variant="danger" onClick={signOut}>
-                  <CareIcon icon="l-sign-out-alt" />
-                  {t("sign_out")}
-                </ButtonV2>
-              </div>
             </div>
           </div>
-          <div className="mt-5 lg:col-span-2 lg:mt-0">
-            {!showEdit && !isLoading && (
-              <div className="m-2 overflow-hidden rounded-lg bg-white px-4 py-5 shadow sm:rounded-lg sm:px-6">
-                <dl className="col-gap-4 row-gap-8 grid grid-cols-1 sm:grid-cols-2">
-                  <div
-                    className="my-2 sm:col-span-1"
-                    id="username-profile-details"
-                  >
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("username")}
-                    </dt>
-                    <dd className="mt-1 text-sm leading-5 text-secondary-900">
-                      {userData?.username || "-"}
-                    </dd>
-                  </div>
-                  <div
-                    className="my-2 sm:col-span-1"
-                    id="contactno-profile-details"
-                  >
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("phone_number")}
-                    </dt>
-                    <dd className="mt-1 text-sm leading-5 text-secondary-900">
-                      {userData?.phone_number || "-"}
-                    </dd>
-                  </div>
+          <div className="flex flex-col gap-2">
+            <ButtonV2
+              onClick={(_) => setShowEdit(!showEdit)}
+              type="button"
+              id="edit-cancel-profile-button"
+            >
+              {showEdit ? t("cancel") : t("edit_user_profile")}
+            </ButtonV2>
+            <ButtonV2 variant="danger" onClick={signOut}>
+              <CareIcon icon="l-sign-out-alt" />
+              {t("sign_out")}
+            </ButtonV2>
+          </div>
+        </div>
+        <div className="mt-5 lg:col-span-2 lg:mt-0">
+          {!showEdit && !isLoading && (
+            <div className="overflow-hidden rounded-lg bg-white px-4 py-5 shadow sm:rounded-lg sm:px-6">
+              <dl className="col-gap-4 row-gap-8 grid grid-cols-1 sm:grid-cols-2">
+                <div
+                  className="my-2 sm:col-span-1"
+                  id="username-profile-details"
+                >
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("username")}
+                  </dt>
+                  <dd className="mt-1 text-sm leading-5 text-secondary-900">
+                    {userData?.username || "-"}
+                  </dd>
+                </div>
+                <div
+                  className="my-2 sm:col-span-1"
+                  id="contactno-profile-details"
+                >
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("phone_number")}
+                  </dt>
+                  <dd className="mt-1 text-sm leading-5 text-secondary-900">
+                    {userData?.phone_number || "-"}
+                  </dd>
+                </div>
 
-                  <div
-                    className="my-2 sm:col-span-1"
-                    id="whatsapp-profile-details"
-                  >
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("whatsapp_number")}
-                    </dt>
-                    <dd className="mt-1 text-sm leading-5 text-secondary-900">
-                      {userData?.alt_phone_number || "-"}
-                    </dd>
-                  </div>
-                  <div
-                    className="my-2 sm:col-span-1"
-                    id="emailid-profile-details"
-                  >
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("email")}
-                    </dt>
-                    <dd className="mt-1 text-sm leading-5 text-secondary-900">
-                      {userData?.email || "-"}
-                    </dd>
-                  </div>
-                  <div
-                    className="my-2 sm:col-span-1"
-                    id="firstname-profile-details"
-                  >
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("first_name")}
-                    </dt>
-                    <dd className="mt-1 text-sm leading-5 text-secondary-900">
-                      {userData?.first_name || "-"}
-                    </dd>
-                  </div>
-                  <div
-                    className="my-2 sm:col-span-1"
-                    id="lastname-profile-details"
-                  >
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("last_name")}
-                    </dt>
-                    <dd className="mt-1 text-sm leading-5 text-secondary-900">
-                      {userData?.last_name || "-"}
-                    </dd>
-                  </div>
-                  <div
-                    className="my-2 sm:col-span-1"
-                    id="date_of_birth-profile-details"
-                  >
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("date_of_birth")}
-                    </dt>
-                    <dd className="mt-1 text-sm leading-5 text-secondary-900">
-                      {userData?.date_of_birth
-                        ? formatDate(userData?.date_of_birth)
+                <div
+                  className="my-2 sm:col-span-1"
+                  id="whatsapp-profile-details"
+                >
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("whatsapp_number")}
+                  </dt>
+                  <dd className="mt-1 text-sm leading-5 text-secondary-900">
+                    {userData?.alt_phone_number || "-"}
+                  </dd>
+                </div>
+                <div
+                  className="my-2 sm:col-span-1"
+                  id="emailid-profile-details"
+                >
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("email")}
+                  </dt>
+                  <dd className="mt-1 text-sm leading-5 text-secondary-900">
+                    {userData?.email || "-"}
+                  </dd>
+                </div>
+                <div
+                  className="my-2 sm:col-span-1"
+                  id="firstname-profile-details"
+                >
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("first_name")}
+                  </dt>
+                  <dd className="mt-1 text-sm leading-5 text-secondary-900">
+                    {userData?.first_name || "-"}
+                  </dd>
+                </div>
+                <div
+                  className="my-2 sm:col-span-1"
+                  id="lastname-profile-details"
+                >
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("last_name")}
+                  </dt>
+                  <dd className="mt-1 text-sm leading-5 text-secondary-900">
+                    {userData?.last_name || "-"}
+                  </dd>
+                </div>
+                <div
+                  className="my-2 sm:col-span-1"
+                  id="date_of_birth-profile-details"
+                >
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("date_of_birth")}
+                  </dt>
+                  <dd className="mt-1 text-sm leading-5 text-secondary-900">
+                    {userData?.date_of_birth
+                      ? formatDate(userData?.date_of_birth)
+                      : "-"}
+                  </dd>
+                </div>
+                <div className="my-2 sm:col-span-1">
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("access_level")}
+                  </dt>
+                  <dd className="badge badge-pill mt-1 bg-primary-500 text-sm text-white">
+                    <CareIcon icon="l-user-check" className="mr-2 text-lg" />
+                    {userData?.user_type || "-"}
+                  </dd>
+                </div>
+                <div className="my-2 sm:col-span-1" id="gender-profile-details">
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("gender")}
+                  </dt>
+                  <dd className="mt-1 text-sm leading-5 text-secondary-900">
+                    {userData?.gender || "-"}
+                  </dd>
+                </div>
+                <div className="my-2 sm:col-span-1">
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("local_body")}
+                  </dt>
+                  <dd className="mt-1 text-sm leading-5 text-secondary-900">
+                    {userData?.local_body_object?.name || "-"}
+                  </dd>
+                </div>
+                <div className="my-2 sm:col-span-1">
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("district")}
+                  </dt>
+                  <dd className="mt-1 text-sm leading-5 text-secondary-900">
+                    {userData?.district_object?.name || "-"}
+                  </dd>
+                </div>
+                <div className="my-2 sm:col-span-1">
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("state")}
+                  </dt>
+                  <dd className="mt-1 text-sm leading-5 text-secondary-900">
+                    {userData?.state_object?.name || "-"}
+                  </dd>
+                </div>
+                <div className="my-2 sm:col-span-1">
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("skills")}
+                  </dt>
+                  <dd className="mt-1 text-sm leading-5 text-secondary-900">
+                    <div
+                      className="flex flex-wrap gap-2"
+                      id="already-linked-skills"
+                    >
+                      {skillsView?.results?.length
+                        ? skillsView.results?.map((skill: SkillModel) => {
+                            return (
+                              <span className="flex items-center gap-2 rounded-full border-secondary-300 bg-secondary-200 px-3 text-xs text-secondary-700">
+                                <p className="py-1.5">
+                                  {skill.skill_object.name}
+                                </p>
+                              </span>
+                            );
+                          })
                         : "-"}
-                    </dd>
-                  </div>
-                  <div className="my-2 sm:col-span-1">
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("access_level")}
-                    </dt>
-                    <dd className="badge badge-pill mt-1 bg-primary-500 text-sm text-white">
-                      <CareIcon icon="l-user-check" className="mr-1 text-lg" />{" "}
-                      {userData?.user_type || "-"}
-                    </dd>
-                  </div>
-                  <div
-                    className="my-2 sm:col-span-1"
-                    id="gender-profile-details"
-                  >
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("gender")}
-                    </dt>
-                    <dd className="mt-1 text-sm leading-5 text-secondary-900">
-                      {userData?.gender || "-"}
-                    </dd>
-                  </div>
-                  <div className="my-2 sm:col-span-1">
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("local_body")}
-                    </dt>
-                    <dd className="mt-1 text-sm leading-5 text-secondary-900">
-                      {userData?.local_body_object?.name || "-"}
-                    </dd>
-                  </div>
-                  <div className="my-2 sm:col-span-1">
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("district")}
-                    </dt>
-                    <dd className="mt-1 text-sm leading-5 text-secondary-900">
-                      {userData?.district_object?.name || "-"}
-                    </dd>
-                  </div>
-                  <div className="my-2 sm:col-span-1">
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("state")}
-                    </dt>
-                    <dd className="mt-1 text-sm leading-5 text-secondary-900">
-                      {userData?.state_object?.name || "-"}
-                    </dd>
-                  </div>
-                  <div className="my-2 sm:col-span-1">
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("skills")}
-                    </dt>
-                    <dd className="mt-1 text-sm leading-5 text-secondary-900">
-                      <div
-                        className="flex flex-wrap gap-2"
-                        id="already-linked-skills"
+                    </div>
+                  </dd>
+                </div>
+                <div
+                  className="my-2 sm:col-span-1"
+                  id="averageworkinghour-profile-details"
+                >
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("average_weekly_working_hours")}
+                  </dt>
+                  <dd className="mt-1 text-sm leading-5 text-secondary-900">
+                    {userData?.weekly_working_hours ?? "-"}
+                  </dd>
+                </div>
+                <div
+                  className="my-2 sm:col-span-2"
+                  id="videoconnectlink-profile-details"
+                >
+                  <dt className="text-sm font-medium leading-5 text-black">
+                    {t("video_conference_link")}
+                  </dt>
+                  <dd className="mt-1 break-words text-sm leading-5 text-secondary-900">
+                    {userData?.video_connect_link ? (
+                      <a
+                        className="text-blue-500"
+                        href={userData?.video_connect_link}
+                        target="_blank"
+                        rel="noreferrer"
                       >
-                        {skillsView?.results?.length
-                          ? skillsView.results?.map((skill: SkillModel) => {
-                              return (
-                                <span className="flex items-center gap-2 rounded-full border-secondary-300 bg-secondary-200 px-3 text-xs text-secondary-700">
-                                  <p className="py-1.5">
-                                    {skill.skill_object.name}
-                                  </p>
-                                </span>
-                              );
-                            })
-                          : "-"}
-                      </div>
-                    </dd>
-                  </div>
-                  <div
-                    className="my-2 sm:col-span-1"
-                    id="averageworkinghour-profile-details"
-                  >
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("average_weekly_working_hours")}
-                    </dt>
-                    <dd className="mt-1 text-sm leading-5 text-secondary-900">
-                      {userData?.weekly_working_hours ?? "-"}
-                    </dd>
-                  </div>
-                  <div
-                    className="my-2 sm:col-span-2"
-                    id="videoconnectlink-profile-details"
-                  >
-                    <dt className="text-sm font-medium leading-5 text-black">
-                      {t("video_conference_link")}
-                    </dt>
-                    <dd className="mt-1 break-words text-sm leading-5 text-secondary-900">
-                      {userData?.video_connect_link ? (
-                        <a
-                          className="text-blue-500"
-                          href={userData?.video_connect_link}
-                          target="_blank"
-                          rel="noreferrer"
-                        >
-                          {userData?.video_connect_link}
-                        </a>
-                      ) : (
-                        "-"
+                        {userData?.video_connect_link}
+                      </a>
+                    ) : (
+                      "-"
+                    )}
+                  </dd>
+                </div>
+              </dl>
+            </div>
+          )}
+          {showEdit && (
+            <div className="space-y-4">
+              <form action="#" method="POST">
+                <div className="shadow sm:rounded-md">
+                  <div className="bg-white px-4 pt-5">
+                    <div className="grid grid-cols-6 gap-4">
+                      <TextFormField
+                        {...fieldProps("firstName")}
+                        required
+                        label={t("first_name")}
+                        className="col-span-6 sm:col-span-3"
+                      />
+                      <TextFormField
+                        {...fieldProps("lastName")}
+                        required
+                        label={t("last_name")}
+                        className="col-span-6 sm:col-span-3"
+                      />
+                      <DateFormField
+                        {...fieldProps("date_of_birth")}
+                        label={t("date_of_birth")}
+                        required
+                        className="col-span-6 sm:col-span-3"
+                        value={getDate(states.form.date_of_birth)}
+                        position="LEFT"
+                        disableFuture={true}
+                      />
+                      <SelectFormField
+                        {...fieldProps("gender")}
+                        label={t("gender")}
+                        className="col-span-6 sm:col-span-3"
+                        required
+                        optionLabel={(o) => o.text}
+                        optionValue={(o) => o.text}
+                        options={GENDER_TYPES}
+                      />
+                      <PhoneNumberFormField
+                        {...fieldProps("phoneNumber")}
+                        label={t("phone_number")}
+                        className="col-span-6 sm:col-span-3"
+                        required
+                        placeholder={t("phone_number")}
+                        types={["mobile", "landline"]}
+                      />
+                      <PhoneNumberFormField
+                        {...fieldProps("altPhoneNumber")}
+                        label={t("whatsapp_number")}
+                        className="col-span-6 sm:col-span-3"
+                        placeholder={t("whatsapp_number")}
+                        types={["mobile"]}
+                      />
+                      <TextFormField
+                        {...fieldProps("email")}
+                        label={t("email")}
+                        className="col-span-6 sm:col-span-3"
+                        required
+                        type="email"
+                      />
+                      {(states.form.user_type === "Doctor" ||
+                        states.form.user_type === "Nurse") && (
+                        <TextFormField
+                          {...fieldProps("qualification")}
+                          required
+                          className="col-span-6 sm:col-span-3"
+                          label={t("qualification")}
+                          placeholder={t("qualification")}
+                        />
                       )}
-                    </dd>
-                  </div>
-                </dl>
-              </div>
-            )}
-            {showEdit && (
-              <div className="space-y-4">
-                <form action="#" method="POST">
-                  <div className="shadow sm:rounded-md">
-                    <div className="bg-white px-4 pt-5">
-                      <div className="grid grid-cols-6 gap-4">
-                        <TextFormField
-                          {...fieldProps("firstName")}
-                          required
-                          label={t("first_name")}
-                          className="col-span-6 sm:col-span-3"
-                        />
-                        <TextFormField
-                          {...fieldProps("lastName")}
-                          required
-                          label={t("last_name")}
-                          className="col-span-6 sm:col-span-3"
-                        />
-                        <DateFormField
-                          {...fieldProps("date_of_birth")}
-                          label={t("date_of_birth")}
-                          required
-                          className="col-span-6 sm:col-span-3"
-                          value={getDate(states.form.date_of_birth)}
-                          position="LEFT"
-                          disableFuture={true}
-                        />
-                        <SelectFormField
-                          {...fieldProps("gender")}
-                          label={t("gender")}
-                          className="col-span-6 sm:col-span-3"
-                          required
-                          optionLabel={(o) => o.text}
-                          optionValue={(o) => o.text}
-                          options={GENDER_TYPES}
-                        />
-                        <PhoneNumberFormField
-                          {...fieldProps("phoneNumber")}
-                          label={t("phone_number")}
-                          className="col-span-6 sm:col-span-3"
-                          required
-                          placeholder={t("phone_number")}
-                          types={["mobile", "landline"]}
-                        />
-                        <PhoneNumberFormField
-                          {...fieldProps("altPhoneNumber")}
-                          label={t("whatsapp_number")}
-                          className="col-span-6 sm:col-span-3"
-                          placeholder={t("whatsapp_number")}
-                          types={["mobile"]}
-                        />
-                        <TextFormField
-                          {...fieldProps("email")}
-                          label={t("email")}
-                          className="col-span-6 sm:col-span-3"
-                          required
-                          type="email"
-                        />
-                        {(states.form.user_type === "Doctor" ||
-                          states.form.user_type === "Nurse") && (
+                      {states.form.user_type === "Doctor" && (
+                        <>
                           <TextFormField
-                            {...fieldProps("qualification")}
+                            {...fieldProps("doctor_experience_commenced_on")}
                             required
                             className="col-span-6 sm:col-span-3"
-                            label={t("qualification")}
-                            placeholder={t("qualification")}
+                            type="number"
+                            min={0}
+                            label={t("years_of_experience")}
+                            placeholder={t("years_of_experience_of_the_doctor")}
                           />
-                        )}
-                        {states.form.user_type === "Doctor" && (
-                          <>
-                            <TextFormField
-                              {...fieldProps("doctor_experience_commenced_on")}
-                              required
-                              className="col-span-6 sm:col-span-3"
-                              type="number"
-                              min={0}
-                              label={t("years_of_experience")}
-                              placeholder={t(
-                                "years_of_experience_of_the_doctor",
-                              )}
-                            />
-                            <TextFormField
-                              {...fieldProps(
-                                "doctor_medical_council_registration",
-                              )}
-                              required
-                              className="col-span-6 sm:col-span-3"
-                              label={t("medical_council_registration")}
-                              placeholder={t(
-                                "doctor_s_medical_council_registration",
-                              )}
-                            />
-                          </>
-                        )}
-                        <TextFormField
-                          {...fieldProps("weekly_working_hours")}
-                          label={t("average_weekly_working_hours")}
-                          className="col-span-6 sm:col-span-3"
-                          type="number"
-                          min={0}
-                          max={168}
-                        />
-                        <TextFormField
-                          {...fieldProps("video_connect_link")}
-                          label={t("video_conference_link")}
-                          className="col-span-6 sm:col-span-6"
-                          type="url"
-                        />
-                      </div>
-                    </div>
-                    <div className="bg-secondary-50 px-4 py-3 text-right sm:px-6">
-                      <Submit onClick={handleSubmit} label={t("update")} />
-                    </div>
-                  </div>
-                </form>
-                <form action="#" method="POST">
-                  <div className="overflow-hidden shadow sm:rounded-md">
-                    <div className="bg-white px-4 pt-5">
-                      <div className="grid grid-cols-6 gap-4">
-                        <TextFormField
-                          name="old_password"
-                          label={t("current_password")}
-                          className="col-span-6 sm:col-span-3"
-                          type="password"
-                          value={changePasswordForm.old_password}
-                          onChange={(e) =>
-                            setChangePasswordForm({
-                              ...changePasswordForm,
-                              old_password: e.value,
-                            })
-                          }
-                          error={changePasswordErrors.old_password}
-                          required
-                        />
-                        <div className="col-span-6 sm:col-span-3">
                           <TextFormField
-                            name="new_password_1"
-                            label={t("new_password")}
-                            type="password"
-                            value={changePasswordForm.new_password_1}
-                            className="peer col-span-6 sm:col-span-3"
-                            onChange={(e) => {
-                              setChangePasswordForm({
-                                ...changePasswordForm,
-                                new_password_1: e.value,
-                              });
-                            }}
+                            {...fieldProps(
+                              "doctor_medical_council_registration",
+                            )}
                             required
+                            className="col-span-6 sm:col-span-3"
+                            label={t("medical_council_registration")}
+                            placeholder={t(
+                              "doctor_s_medical_council_registration",
+                            )}
                           />
-                          <div className="text-small mb-2 hidden pl-2 text-secondary-500 peer-focus-within:block">
-                            {validateRule(
-                              changePasswordForm.new_password_1?.length >= 8,
-                              "Password should be atleast 8 characters long",
-                            )}
-                            {validateRule(
-                              changePasswordForm.new_password_1 !==
-                                changePasswordForm.new_password_1.toUpperCase(),
-                              "Password should contain at least 1 lowercase letter",
-                            )}
-                            {validateRule(
-                              changePasswordForm.new_password_1 !==
-                                changePasswordForm.new_password_1.toLowerCase(),
-                              "Password should contain at least 1 uppercase letter",
-                            )}
-                            {validateRule(
-                              /\d/.test(changePasswordForm.new_password_1),
-                              "Password should contain at least 1 number",
-                            )}
-                          </div>
-                        </div>
-                        <div className="col-span-6 sm:col-span-3">
-                          <TextFormField
-                            name="new_password_2"
-                            label={t("new_password_confirmation")}
-                            className="peer col-span-6 sm:col-span-3"
-                            type="password"
-                            value={changePasswordForm.new_password_2}
-                            onChange={(e) => {
-                              setChangePasswordForm({
-                                ...changePasswordForm,
-                                new_password_2: e.value,
-                              });
-                            }}
-                          />
-                          {changePasswordForm.new_password_2.length > 0 && (
-                            <div className="text-small mb-2 hidden pl-2 text-secondary-500 peer-focus-within:block">
-                              {validateRule(
-                                changePasswordForm.new_password_1 ===
-                                  changePasswordForm.new_password_2,
-                                "Confirm password should match the new password",
-                              )}
-                            </div>
-                          )}
-                        </div>
-                      </div>
-                    </div>
-                    <div className="bg-secondary-50 px-4 py-3 text-right sm:px-6">
-                      <Submit
-                        onClick={changePassword}
-                        label={t("change_password")}
+                        </>
+                      )}
+                      <TextFormField
+                        {...fieldProps("weekly_working_hours")}
+                        label={t("average_weekly_working_hours")}
+                        className="col-span-6 sm:col-span-3"
+                        type="number"
+                        min={0}
+                        max={168}
+                      />
+                      <TextFormField
+                        {...fieldProps("video_connect_link")}
+                        label={t("video_conference_link")}
+                        className="col-span-6 sm:col-span-6"
+                        type="url"
                       />
                     </div>
                   </div>
-                </form>
-              </div>
-            )}
-          </div>
-        </div>
-
-        <div className="mb-8 mt-6 md:grid md:grid-cols-3 md:gap-6">
-          <div className="md:col-span-1">
-            <div className="px-4 sm:px-0">
-              <h3 className="text-lg font-medium leading-6 text-secondary-900">
-                {t("language_selection")}
-              </h3>
-              <p className="mt-1 text-sm leading-5 text-secondary-600">
-                {t("set_your_local_language")}
-              </p>
-            </div>
-          </div>
-          <div className="mt-5 md:col-span-2 md:mt-0">
-            <LanguageSelector className="w-full bg-white" />
-          </div>
-        </div>
-        <div className="mb-8 mt-6 md:grid md:grid-cols-3 md:gap-6">
-          <div className="md:col-span-1">
-            <div className="px-4 sm:px-0">
-              <h3 className="text-lg font-medium leading-6 text-secondary-900">
-                {t("software_update")}
-              </h3>
-              <p className="mt-1 text-sm leading-5 text-secondary-600">
-                {t("check_for_available_update")}
-              </p>
-            </div>
-          </div>
-          {updateStatus.isUpdateAvailable && (
-            <UpdatableApp silentlyAutoUpdate={false}>
-              <ButtonV2 disabled={true}>
-                <div className="flex items-center gap-4">
-                  <CareIcon icon="l-exclamation" className="text-2xl" />
-                  {t("update_available")}
+                  <div className="bg-secondary-50 px-4 py-3 text-right sm:px-6">
+                    <Submit onClick={handleSubmit} label={t("update")} />
+                  </div>
                 </div>
-              </ButtonV2>
-            </UpdatableApp>
+              </form>
+              <form action="#" method="POST">
+                <div className="overflow-hidden shadow sm:rounded-md">
+                  <div className="bg-white px-4 pt-5">
+                    <div className="grid grid-cols-6 gap-4">
+                      <TextFormField
+                        name="old_password"
+                        label={t("current_password")}
+                        className="col-span-6 sm:col-span-3"
+                        type="password"
+                        value={changePasswordForm.old_password}
+                        onChange={(e) =>
+                          setChangePasswordForm({
+                            ...changePasswordForm,
+                            old_password: e.value,
+                          })
+                        }
+                        error={changePasswordErrors.old_password}
+                        required
+                      />
+                      <div className="col-span-6 sm:col-span-3">
+                        <TextFormField
+                          name="new_password_1"
+                          label={t("new_password")}
+                          type="password"
+                          value={changePasswordForm.new_password_1}
+                          className="peer col-span-6 sm:col-span-3"
+                          onChange={(e) => {
+                            setChangePasswordForm({
+                              ...changePasswordForm,
+                              new_password_1: e.value,
+                            });
+                          }}
+                          required
+                        />
+                        <div className="text-small mb-2 hidden pl-2 text-secondary-500 peer-focus-within:block">
+                          {validateRule(
+                            changePasswordForm.new_password_1?.length >= 8,
+                            "Password should be atleast 8 characters long",
+                          )}
+                          {validateRule(
+                            changePasswordForm.new_password_1 !==
+                              changePasswordForm.new_password_1.toUpperCase(),
+                            "Password should contain at least 1 lowercase letter",
+                          )}
+                          {validateRule(
+                            changePasswordForm.new_password_1 !==
+                              changePasswordForm.new_password_1.toLowerCase(),
+                            "Password should contain at least 1 uppercase letter",
+                          )}
+                          {validateRule(
+                            /\d/.test(changePasswordForm.new_password_1),
+                            "Password should contain at least 1 number",
+                          )}
+                        </div>
+                      </div>
+                      <div className="col-span-6 sm:col-span-3">
+                        <TextFormField
+                          name="new_password_2"
+                          label={t("new_password_confirmation")}
+                          className="peer col-span-6 sm:col-span-3"
+                          type="password"
+                          value={changePasswordForm.new_password_2}
+                          onChange={(e) => {
+                            setChangePasswordForm({
+                              ...changePasswordForm,
+                              new_password_2: e.value,
+                            });
+                          }}
+                        />
+                        {changePasswordForm.new_password_2.length > 0 && (
+                          <div className="text-small mb-2 hidden pl-2 text-secondary-500 peer-focus-within:block">
+                            {validateRule(
+                              changePasswordForm.new_password_1 ===
+                                changePasswordForm.new_password_2,
+                              "Confirm password should match the new password",
+                            )}
+                          </div>
+                        )}
+                      </div>
+                    </div>
+                  </div>
+                  <div className="bg-secondary-50 px-4 py-3 text-right sm:px-6">
+                    <Submit
+                      onClick={changePassword}
+                      label={t("change_password")}
+                    />
+                  </div>
+                </div>
+              </form>
+            </div>
           )}
-          <div className="mt-5 md:col-span-2 md:mt-0">
-            {!updateStatus.isUpdateAvailable && (
-              <ButtonV2
-                disabled={updateStatus.isChecking}
-                onClick={checkUpdates}
-              >
-                {" "}
-                <div className="flex items-center gap-4">
-                  <CareIcon
-                    icon="l-sync"
-                    className={classNames(
-                      "text-2xl",
-                      updateStatus.isChecking && "animate-spin",
-                    )}
-                  />
-                  {updateStatus.isChecking
-                    ? t("checking_for_update")
-                    : t("check_for_update")}
-                </div>
-              </ButtonV2>
-            )}
-          </div>
         </div>
       </div>
-    </div>
+
+      <div className="mb-8 mt-6 md:grid md:grid-cols-3 md:gap-6">
+        <div className="md:col-span-1">
+          <div className="px-4 sm:px-0">
+            <h3 className="text-lg font-medium leading-6 text-secondary-900">
+              {t("language_selection")}
+            </h3>
+            <p className="mt-1 text-sm leading-5 text-secondary-600">
+              {t("set_your_local_language")}
+            </p>
+          </div>
+        </div>
+        <div className="mt-5 md:col-span-2 md:mt-0">
+          <LanguageSelector className="w-full bg-white" />
+        </div>
+      </div>
+      <div className="mb-8 mt-6 md:grid md:grid-cols-3 md:gap-6">
+        <div className="md:col-span-1">
+          <div className="px-4 sm:px-0">
+            <h3 className="text-lg font-medium leading-6 text-secondary-900">
+              {t("software_update")}
+            </h3>
+            <p className="mt-1 text-sm leading-5 text-secondary-600">
+              {t("check_for_available_update")}
+            </p>
+          </div>
+        </div>
+        {updateStatus.isUpdateAvailable && (
+          <UpdatableApp silentlyAutoUpdate={false}>
+            <ButtonV2 disabled={true}>
+              <div className="flex items-center gap-4">
+                <CareIcon icon="l-exclamation" className="text-2xl" />
+                {t("update_available")}
+              </div>
+            </ButtonV2>
+          </UpdatableApp>
+        )}
+        <div className="mt-5 md:col-span-2 md:mt-0">
+          {!updateStatus.isUpdateAvailable && (
+            <ButtonV2 disabled={updateStatus.isChecking} onClick={checkUpdates}>
+              {" "}
+              <div className="flex items-center gap-4">
+                <CareIcon
+                  icon="l-sync"
+                  className={classNames(
+                    "text-2xl",
+                    updateStatus.isChecking && "animate-spin",
+                  )}
+                />
+                {updateStatus.isChecking
+                  ? t("checking_for_update")
+                  : t("check_for_update")}
+              </div>
+            </ButtonV2>
+          )}
+        </div>
+      </div>
+    </Page>
   );
 }
