+++ src/Locale/en.json
@@ -569,6 +569,7 @@
   "duplicate_patient_record_confirmation": "Admit the patient record to your facility by adding the year of birth",
   "duplicate_patient_record_rejection": "I confirm that the suspect / patient I want to create is not on the list.",
   "edit": "Edit",
+  "edit_avatar": "Edit Avatar",
   "edit_caution_note": "A new prescription will be added to the consultation with the edited details and the current prescription will be discontinued.",
   "edit_cover_photo": "Edit Cover Photo",
   "edit_history": "Edit History",
@@ -843,12 +844,12 @@
   "no_changes": "No changes",
   "no_changes_made": "No changes made",
   "no_consultation_updates": "No consultation updates",
-  "no_cover_photo_uploaded_for_this_facility": "No cover photo uploaded for this facility",
   "no_data_found": "No data found",
   "no_duplicate_facility": "You should not create duplicate facilities",
   "no_facilities": "No Facilities found",
   "no_files_found": "No {{type}} files found",
   "no_home_facility": "No home facility assigned",
+  "no_image_found": "No image found",
   "no_investigation": "No investigation Reports found",
   "no_investigation_suggestions": "No Investigation Suggestions",
   "no_linked_facilities": "No Linked Facilities",
