+++ src/Locale/hi.json
@@ -506,7 +506,6 @@
   "no_changes": "कोई परिवर्तन नहीं",
   "no_changes_made": "कोई परिवर्तन नहीं किया गया",
   "no_consultation_updates": "कोई परामर्श अपडेट नहीं",
-  "no_cover_photo_uploaded_for_this_facility": "इस सुविधा के लिए कोई कवर फ़ोटो अपलोड नहीं किया गया",
   "no_data_found": "डाटा प्राप्त नहीं हुआ",
   "no_duplicate_facility": "आपको डुप्लिकेट सुविधाएं नहीं बनानी चाहिए",
   "no_facilities": "कोई सुविधा नहीं मिली",
