+++ src/Locale/ml.json
@@ -506,7 +506,6 @@
   "no_changes": "മാറ്റങ്ങളൊന്നുമില്ല",
   "no_changes_made": "മാറ്റങ്ങളൊന്നും വരുത്തിയിട്ടില്ല",
   "no_consultation_updates": "കൺസൾട്ടേഷൻ അപ്‌ഡേറ്റുകളൊന്നുമില്ല",
-  "no_cover_photo_uploaded_for_this_facility": "ഈ സൗകര്യത്തിനായി കവർ ഫോട്ടോ അപ്‌ലോഡ് ചെയ്‌തിട്ടില്ല",
   "no_data_found": "വിവരങ്ങളൊന്നും കണ്ടെത്തിയില്ല",
   "no_duplicate_facility": "അനധികൃതമായി ഫെസിലിറ്റികള്‍ സൃഷ്ടിക്കരുത്",
   "no_facilities": "ഫെസിലിറ്റികളൊന്നും കണ്ടെത്തുവാനായില്ല",
