+++ src/Locale/ta.json
@@ -506,7 +506,6 @@
   "no_changes": "மாற்றங்கள் இல்லை",
   "no_changes_made": "எந்த மாற்றமும் செய்யப்படவில்லை",
   "no_consultation_updates": "ஆலோசனை அறிவிப்புகள் இல்லை",
-  "no_cover_photo_uploaded_for_this_facility": "இந்த வசதிக்காக அட்டைப் புகைப்படம் பதிவேற்றப்படவில்லை",
   "no_data_found": "தரவு எதுவும் கிடைக்கவில்லை",
   "no_duplicate_facility": "நீங்கள் நகல் வசதிகளை உருவாக்கக்கூடாது",
   "no_facilities": "வசதிகள் எதுவும் கிடைக்கவில்லை",
